Prefer "3D" over "3d" in method and parameter names
As "3D" is a word, according to .NET docs regarding capitalization
conventions (https://docs.microsoft.com/en-us/dotnet/standard/design-guidelines/capitalization-conventions)
"3D" could be correctly used as "3d" in method and parameter names.
However as "3D" can also be commonly thought of as simply an acronym,
it would also be correct to represent it as "3D" in method and
parameter names. We found an example of a Microsoft API which uses
"3D" instead of "3d", so we're going with that.